# [FIX] stock: speedup inventory lines creation