--- a/addons/stock/tests/test_inventory.py
+++ b/addons/stock/tests/test_inventory.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-from odoo.exceptions import ValidationError
+from odoo.exceptions import ValidationError, UserError
 from odoo.tests.common import Form, SavepointCase
 
 
@@ -802,3 +802,24 @@ class TestInventory(SavepointCase):
         self.assertEqual(len(line_ids_loc2), 1)
         self.assertEqual(line_ids_loc1[0].theoretical_qty, 10)
         self.assertEqual(line_ids_loc2[0].theoretical_qty, 0)
+
+    def test_inventory_line_duplicates(self):
+        """ Checks that creating duplicated inventory lines
+        raises a UserError.
+        """
+        inventory = self.env['stock.inventory'].create({
+            'name': 'Existing Inventory',
+            'exhausted': True,
+            'location_ids': [(4, self.stock_location.id)],
+            'product_ids': [(4, self.product1.id)]
+        })
+        inventory.action_start()
+
+        dup_vals = [{
+            'inventory_id': inventory.id,
+            'product_id': self.product1.id,
+            'location_id': self.stock_location.id,
+        }]
+
+        with self.assertRaises(UserError), self.cr.savepoint():
+            self.env['stock.inventory.line'].create(dup_vals)