Browse product.product outside of for loop in
`inventory_line.create` to speedup 
the `product.product_tmpl_id.uom_id.id` line.

Also, replace `product.uom_id.id` by `product.product_tmpl_id.uom_id.id`
to avoid call to `compute_related` inside the for loop.

Replace search_count in for loop of
`inventory_line._check_no_duplicate_line` by a read_group call.

Add a small test for `_check_no_duplicate_line` just to be
safe.

#### speedup

Testing database with growing number of products.

##### browse speedup

| Number of Products | Before PR | After PR |
|:---------------------:|:----------:|:---------:|
| 1000 | 5.5s | 50ms |
| 2500 | 15s | 109ms|
| 5000 | 29s | 220ms|
| 10000 | 50s | 400ms|

##### check_no_duplicate_line speedup

| Number of Products | Before PR | After PR |
|:---------------------:|:----------:|:---------:|
| 1000 | 1.5s | 200ms |
| 2500 | 3s | 467ms|
| 5000 | 13.5s | 750ms|
| 10000 | 41s | 1.6s|

Inventory with 10k products, `action_start` time: 1m40s -> 9s.

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr