@robodoo 2021-11-09
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/79587).