@Aurelienvd 2021-11-16
> 
> 
> Would be cool to test/re-write it against V13 too though :) Now it only fixes a few versions.

@Yenthe666 I'll do it later this week.